refactor(resources): move to new @linktank/links lib
