+++ apps/api/src/app/app.module.ts
@@ -4,11 +4,11 @@ import { HttpModule } from '@nestjs/axios';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 
-import { LinksModule, CertificatesModule } from '@linktank/resources';
+import { LinksModule } from '@linktank/links';
 import { HealthModule } from '@linktank/health';
 
 @Module({
-  imports: [HttpModule, LinksModule, CertificatesModule, HealthModule],
+  imports: [HttpModule, LinksModule, HealthModule],
   controllers: [AppController],
   providers: [AppService],
 })
