+++ libs/links/README.md
@@ -1,0 +1,7 @@
+# links
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test links` to execute the unit tests via [Jest](https://jestjs.io).
