+++ libs/links/src/index.ts
@@ -1,11 +1,6 @@
 export * from './links/links.module';
-export * from './certificates/certificates.module';
-
 export * from './links/links.service';
-export * from './certificates/certificates.service';
-
 export * from './links/links.health';
-// export * from './certificates/certificates.health';
 
 export * from './links/dto/create-link.dto';
 export * from './links/dto/update-link.dto';
