--- libs/resources/README.md
@@ -1,7 +1,0 @@
-# resources
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test resources` to execute the unit tests via [Jest](https://jestjs.io).
