--- libs/resources/src/certificates/certificates.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { CertificatesController } from './certificates.controller';
-import { CertificatesService } from './certificates.service';
-
-describe('CertificatesController', () => {
-  let controller: CertificatesController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [CertificatesController],
-      providers: [CertificatesService],
-    }).compile();
-
-    controller = module.get<CertificatesController>(CertificatesController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
