--- libs/resources/src/certificates/certificates.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { CertificatesService } from './certificates.service';
-import { CertificatesController } from './certificates.controller';
-
-@Module({
-  controllers: [CertificatesController],
-  providers: [CertificatesService],
-  exports: [CertificatesService],
-})
-export class CertificatesModule {}
