--- libs/resources/src/certificates/dto/create-certificate.dto.ts
@@ -1,1 +1,0 @@
-export class CreateCertificateDto {}
