--- libs/resources/src/certificates/dto/update-certificate.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateCertificateDto } from './create-certificate.dto';
-
-export class UpdateCertificateDto extends PartialType(CreateCertificateDto) {}
