--- libs/resources/src/certificates/entities/certificate.entity.ts
@@ -1,1 +1,0 @@
-export class Certificate {}
