+++ tsconfig.base.json
@@ -16,7 +16,7 @@
     "baseUrl": ".",
     "paths": {
       "@linktank/health": ["libs/health/src/index.ts"],
-      "@linktank/resources": ["libs/resources/src/index.ts"],
+      "@linktank/links": ["libs/links/src/index.ts"],
       "@linktank/ui": ["libs/ui/src/index.ts"]
     }
   },
